+++ ios/flutter_eco_mode/Sources/flutter-eco-mode/listeners/BatteryStateListener.swift
@@ -39,8 +39,7 @@ class BatteryStateListener: BatteryStateStreamHandler, DisposableStreamListener 
     func dispose(binaryMessenger: FlutterBinaryMessenger) {
         cleanUp()
         
-        let channelName = "dev.flutter.pigeon.flutter_eco_mode.EcoModeEventChannel.batteryState"
-        FlutterEventChannel(name: channelName, binaryMessenger: binaryMessenger, codec: messagesPigeonMethodCodec).setStreamHandler(nil)
+        FlutterEventChannel(name: EcoModeEventChannels.batteryState, binaryMessenger: binaryMessenger, codec: messagesPigeonMethodCodec).setStreamHandler(nil)
     }
     
     // MARK: - Private methods